Moved process markdownlint results to a dedicated file 🚓
