+++ .markdownlint.json
@@ -1,0 +1,3 @@
+{
+  "extends": "./node_modules/lint-pilot/markdownlint.json"
+}
